Local OCR & Dynamic RAG System/Linkdin-Post-Generator/requirements.txt
fastapi
uvicorn
requests
python-dotenv
pydantic
streamlit
langchain
openai
requests
python-dotenv
groq
